adjust landing
